requirements.txt
asgiref==3.2.10
dataclasses==0.7
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
dnspython==2.0.0
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
pymongo==3.10.1
pytz==2020.1
sqlparse==0.2.4
whitenoise==5.1.0
